Move perft to move_list tests
Moved perft related tests to move_list's tests
