+++ game/src/game/move_list/tests.rs
@@ -1,12 +1,142 @@
-pub mod perft;
+use perft::{count_leaves, perft_leaves, PerftValues};
+use position::prelude::Position;
 
-use perft::*;
-use position::prelude::*;
+mod perft {
+    use position::prelude::{ChessMoveHint, Position};
+
+    use crate::game::move_list::MoveList;
+
+    #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+    pub struct PerftValues {
+        pub nodes: u128,
+        pub captures: u128,
+        pub ep: u128,
+        pub castles: u128,
+        pub promotions: u128,
+        pub checks: u128,
+        pub checkmates: u128,
+    }
+
+    impl PerftValues {
+        pub fn new() -> Self {
+            Self {
+                nodes: 0,
+                captures: 0,
+                ep: 0,
+                castles: 0,
+                promotions: 0,
+                checks: 0,
+                checkmates: 0,
+            }
+        }
+        pub fn add_move(&mut self, hint: ChessMoveHint) {
+            self.nodes += 1;
+            if hint.is_capture() {
+                self.captures += 1;
+            }
+            if hint == ChessMoveHint::EnPassantCapture {
+                self.ep += 1;
+            }
+            if hint == ChessMoveHint::KingCastle || hint == ChessMoveHint::QueenCastle {
+                self.castles += 1;
+            }
+            if hint.is_promotion() {
+                self.promotions += 1;
+            }
+        }
+        pub fn add_check(&mut self) {
+            self.checks += 1;
+        }
+        pub fn add_checkmate(&mut self) {
+            self.checkmates += 1;
+        }
+    }
+
+    pub fn perft_leaves(mut position: Position, depth: u8) -> PerftValues {
+        let mut data = PerftValues::new();
+        collect_perft(&mut position, &mut MoveList::empty(), &mut data, depth - 1);
+        data
+    }
+    pub fn count_leaves(mut position: Position, depth: u8) -> u128 {
+        count_legal_moves(&mut position, &mut MoveList::empty(), depth - 1)
+    }
+
+    fn can_move(position: &mut Position, move_list: &mut MoveList) -> bool {
+        move_list.generate_moves(position);
+
+        while let Some(next_move) = move_list.pop_move() {
+            let unmove = position.make_move(next_move);
+
+            if !position.was_check_ignored() {
+                position.unmake_move(unmove);
+                move_list.pop_group();
+                return true;
+            }
+
+            position.unmake_move(unmove);
+        }
+
+        move_list.pop_group();
+        false
+    }
+
+    fn count_legal_moves(position: &mut Position, move_list: &mut MoveList, depth: u8) -> u128 {
+        let mut count = 0u128;
+        move_list.generate_moves(position);
+
+        while let Some(next_move) = move_list.pop_move() {
+            let unmove = position.make_move(next_move);
+
+            if !position.was_check_ignored() {
+                if depth == 0 {
+                    count += 1;
+                } else {
+                    count += count_legal_moves(position, move_list, depth - 1);
+                }
+            }
+
+            position.unmake_move(unmove);
+        }
+
+        move_list.pop_group();
+        count
+    }
+
+    fn collect_perft(
+        position: &mut Position,
+        move_list: &mut MoveList,
+        data: &mut PerftValues,
+        depth: u8,
+    ) {
+        move_list.generate_moves(position);
+
+        while let Some(next_move) = move_list.pop_move() {
+            let unmove = position.make_move(next_move);
+
+            if !position.was_check_ignored() {
+                if depth == 0 {
+                    data.add_move(next_move.hint);
+                    if position.is_check() {
+                        data.add_check();
+                        if !can_move(position, move_list) {
+                            data.add_checkmate();
+                        }
+                    }
+                } else {
+                    collect_perft(position, move_list, data, depth - 1);
+                }
+            }
+
+            position.unmake_move(unmove);
+        }
+
+        move_list.pop_group();
+    }
+}
 
 const INITIAL_FEN: &str = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
 const KIWIPETE_FEN: &str = "r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R w KQkq - 0 1";
 const CPW_DEBUG_3_FEN: &str = "8/2p5/3p4/KP5r/1R3p1k/8/4P1P1/8 w - - 0 1";
-
 const CPW_MIRROR_W_FEN: &str = "r3k2r/Pppp1ppp/1b3nbN/nP6/BBP1P3/q4N2/Pp1P2PP/R2Q1RK1 w kq - 0 1";
 const CPW_MIRROR_B_FEN: &str = "r2q1rk1/pP1p2pp/Q4n2/bbp1p3/Np6/1B3NBn/pPPP1PPP/R3K2R b KQ - 0 1";
 
@@ -16,7 +146,7 @@ fn perft(fen: &str, depth: u8, expected: PerftValues) {
     let position = Position::try_from_fen(fen).unwrap();
     let values = perft_leaves(position, depth);
     println!("explicit => {values:?}");
-    assert_eq!(values.to_perft_values(), expected);
+    assert_eq!(values, expected);
 }
 
 fn test_move_count(fen: &str, depth: u8, expected: u128) {
