+++ game/src/lib.rs
@@ -6,6 +6,3 @@
 pub mod game;
 pub mod search;
 pub mod transposition_table;
-
-#[cfg(test)]
-mod testing;
